requirements.prod.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.prod.in
#
click==8.0.3
    # via pip-tools
gitdb==4.0.9
    # via gitpython
gitpython==3.1.24
    # via -r requirements.prod.in
pep517==0.12.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements.prod.in
pyyaml==6.0
    # via -r requirements.prod.in
smmap==5.0.0
    # via gitdb
tomli==1.2.2
    # via pep517
typing-extensions==4.0.1
    # via gitpython
wheel==0.37.0
    # via
    #   -r requirements.prod.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
